requirements.txt
streamlit
diffusers
transformers
accelerate
torch
pillow
requests
huggingface_hub  